requirements.txt
Flask
gspread
google-auth
google-auth-oauthlib
google-generativeai
python-dotenv